# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pokemon Anil API",
    "version": "2.0.1",
    "description": "API propria do Pokemon Anil/PT-BR. Dados locais Azul 4.0.6 (interna 4.0.3). Sem consultas externas nos catalogos de jogo."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/moves/type/{type}": {
      "get": {
        "summary": "GET /moves/type/{type}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/types": {
      "get": {
        "summary": "GET /types",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
    "/types/{name}": {
      "get": {
        "summary": "GET /types/{name}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/types/matchup/{atk}/{def}": {
      "get": {
        "summary": "GET /types/matchup/{atk}/{def}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "atk",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "def",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "writeKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}